move daily forecast processing out of getDOM
